--- docs/src/markdown/ListItemDoc.mdx
@@ -1,16 +1,0 @@
-import { ImageGrid } from '../images/Image.tsx';
-import DrawerImg from './drawerAnatomy.png';
-
-# List Item Tag
-
-`<ListItemTag>` is a text item with a colored background and rounded corners that is used to tag lists.
-
-<ImageGrid width="50%" src={DrawerImg} />
-
-## Usage
-
-```tsx
-import { ListItemTag } from '@brightlayer-ui/react-components';
-
-<ListItemTag label={'Foo Bar'} backgroundColor={'white'} fontColor={'#1e90ff'} />;
-```
